OnnxStack.Converter/PixelArtSigma/requirements.txt
diffusers==0.30.2
transformers==4.44.2
torch==2.4.1
peft==0.12.0
tabulate==0.9.0
accelerate==0.34.2
protobuf==3.20.3
onnx==1.16.1
onnxruntime==1.19.2
onnxruntime-directml==1.19.2
onnxruntime_extensions==0.12.0
olive-ai[directml]==0.6.2
